+++ internal/testutil/fakemongo/onboarding.go
@@ -210,12 +210,12 @@ func (r *VerificationConfigRepository) ready(ctx context.Context) error {
 
 var _ ports.VerificationConfigRepository = (*VerificationConfigRepository)(nil)
 
-func (r *AccountAgeConfigRepository) SaveAccountAgeRequirement(ctx context.Context, guildID string, requiredSeconds int64) (domain.AccountAgeConfig, error) {
+func (r *AccountAgeConfigRepository) SaveAccountAgeRequirement(ctx context.Context, guildID string, requiredSeconds float64) (domain.AccountAgeConfig, error) {
 	if err := r.ready(ctx); err != nil {
 		return domain.AccountAgeConfig{}, err
 	}
 	guildID = strings.TrimSpace(guildID)
-	config := domain.AccountAgeConfig{GuildID: guildID, RequiredSeconds: float64(requiredSeconds)}
+	config := domain.AccountAgeConfig{GuildID: guildID, RequiredSeconds: requiredSeconds}
 	if err := config.Validate(); err != nil {
 		return domain.AccountAgeConfig{}, err
 	}
